refact: adapt folders to new architecture, remove old folders
