+++ src/assets/index.js
@@ -1,0 +1,6 @@
+import imag1 from './DSC_0182.png';
+import imag2 from './H21.png';
+import imag3 from './H61.png';
+import partnersLogo from './partners.png';
+
+export { imag1, imag2, imag3, partnersLogo };
