+++ src/assets/partners.png
