--- src/component/layout/ButtonPartners.js
@@ -1,11 +1,0 @@
-import { Button } from '../layout/styles';
-
-function ButtonPartners() {
-  return (
-    <>
-      <Button>Seja nosso parceiro</Button>
-    </>
-  );
-}
-
-export default ButtonPartners;
