--- src/component/layout/PhotoPartners.js
@@ -1,13 +1,0 @@
-import Partners from './../../img/partners.png';
-
-import { PartnersPhoto } from './styles';
-
-function PhotoPartners() {
-  return (
-    <>
-      <PartnersPhoto src={Partners} alt="Nossos parceiros" />
-    </>
-  );
-}
-
-export default PhotoPartners;
