--- src/component/pages/AboutUs.js
@@ -1,16 +1,0 @@
-import Photos from '../layout/Photos';
-import WrapperTextAbout from '../layout/WrapperTextAbout';
-import { Div } from '../layout/styles';
-
-function AboutUs() {
-  return (
-    <>
-      <Div>
-        <Photos />
-        <WrapperTextAbout />
-      </Div>
-    </>
-  );
-}
-
-export default AboutUs;
