--- src/component/pages/WhereAreWe.js
@@ -1,18 +1,0 @@
-import { DivWhere, MapImag } from '../layout/styles';
-
-import WrapperTextWhere from '../layout/WrapperTextWhere';
-
-function WhereAreWe() {
-  return (
-    <>
-      <DivWhere>
-        {/*  <WrapperTextWhere/> ERRO RESOLVIDO*/}
-
-        <WrapperTextWhere />
-        <MapImag />
-      </DivWhere>
-    </>
-  );
-}
-
-export default WhereAreWe;
