+++ src/pages/Home/Home.jsx
@@ -1,0 +1,15 @@
+import { AboutUs, WhereAreWe, Partners } from './components/index';
+
+function Home() {
+  return (
+    <>
+      <AboutUs />
+
+      <WhereAreWe />
+
+      <Partners />
+    </>
+  );
+}
+
+export default Home;
