+++ src/pages/Home/components/AboutUs/AboutUs.jsx
@@ -1,0 +1,15 @@
+import { Container } from './AboutUs.style';
+import { Photos, WrapperTextAbout } from './components';
+
+function AboutUs() {
+  return (
+    <>
+      <Container>
+        <Photos />
+        <WrapperTextAbout />
+      </Container>
+    </>
+  );
+}
+
+export default AboutUs;
