+++ src/pages/Home/components/AboutUs/AboutUs.style.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  position: absolute;
+  width: 1176px;
+  height: 662px;
+  left: 136px;
+  top: 100px;
+`;
