+++ src/pages/Home/components/AboutUs/components/Photos/Photos.styles.js
@@ -1,0 +1,87 @@
+import styled from 'styled-components';
+
+/*DSC_0182 */
+export const Img1 = styled.img`
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  width: 100%;
+  max-width: 609px;
+
+  position: absolute;
+  width: 617px;
+  height: 374px;
+  left: 136px;
+  top: 388px;
+`;
+
+/*Frame the photo DSC_0182  */
+export const FrameImg1 = styled.img`
+  position: absolute;
+  width: 609px;
+  height: 366px;
+  left: 125px; /*136px*/
+  top: 404px; /*396px */
+
+  /* Cores primárias/Principal */
+
+  border: 2px solid #0095f6;
+  box-sizing: border-box;
+  border-radius: 12px;
+`;
+
+/*H21 */
+export const Img2 = styled.img`
+  display: block;
+  margin-right: auto;
+  width: 100%;
+  max-width: 189px;
+
+  position: absolute;
+  width: 197px;
+  height: 280px;
+  left: 136px;
+  top: 100px;
+`;
+/*Frame the photo H21 */
+export const FrameImg2 = styled.img`
+  position: absolute;
+  width: 189px;
+  height: 272px;
+  left: 125px; /* 136px*/
+  top: 90px;
+
+  /* Cores primárias/Principal */
+
+  border: 2px solid #0095f6;
+  box-sizing: border-box;
+  border-radius: 12px;
+`;
+
+/*H61 */
+export const Img3 = styled.img`
+  display: block;
+  margin-right: 100vh;
+  width: 100%;
+  max-width: 420px;
+  top: 100px;
+
+  position: absolute;
+  height: 280px;
+  left: 330px;
+  top: 100px;
+`;
+/*Frame the photo H61 */
+export const FrameImg3 = styled.img`
+  position: absolute;
+  width: 412px;
+  height: 272px;
+  left: 349px;
+  top: 90px;
+
+  /* Cores primárias/Principal */
+
+  border: 2px solid #0095f6;
+  box-sizing: border-box;
+  border-radius: 12px;
+`;
