+++ src/pages/Home/components/AboutUs/components/index.js
@@ -1,0 +1,4 @@
+import Photos from './Photos/Photos';
+import WrapperTextAbout from './WrapperTextAbout/WrapperTextAbout';
+
+export { Photos, WrapperTextAbout };
