+++ src/pages/Home/components/Partners/Partners.jsx
@@ -1,0 +1,20 @@
+import { ButtonPartners, PhotoPartners } from './components';
+import { PartnersText, SpanColor, DivPartners } from './Partners.style';
+
+function Partners() {
+  return (
+    <DivPartners>
+      <PartnersText>
+        <p>
+          Empresas que nos <SpanColor>apoiam</SpanColor>
+        </p>
+      </PartnersText>
+
+      <PhotoPartners />
+
+      <ButtonPartners />
+    </DivPartners>
+  );
+}
+
+export default Partners;
