+++ src/pages/Home/components/Partners/Partners.style.js
@@ -1,0 +1,34 @@
+import styled from 'styled-components';
+
+export const PartnersText = styled.p`
+  position: absolute;
+  width: 379px;
+  height: 42px;
+  left: 530px;
+  top: 1800px;
+  /* Headers/H3 */
+  font-family: Poppins;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 28px;
+  line-height: 42px;
+  /* identical to box height */
+  color: #e9e9e9;
+`;
+
+export const SpanColor = styled.text`
+  color: #0095f6;
+`;
+
+export const DivPartners = styled.menu`
+  position: static;
+  width: 1439px;
+  height: 458px;
+  left: 0px;
+  top: 3400px; /*3400px */
+  /* Inside auto layout */
+  flex: none;
+  order: 5;
+  flex-grow: 0;
+  margin: 16px 0px;
+`;
