+++ src/pages/Home/components/Partners/components/ButtonPartners/ButtonPartners.jsx
@@ -1,0 +1,7 @@
+import { Button } from './ButtonPartners.style';
+
+function ButtonPartners() {
+  return <Button>Seja nosso parceiro</Button>;
+}
+
+export default ButtonPartners;
