+++ src/pages/Home/components/Partners/components/PhotoPartners/PhotoPartners.jsx
@@ -1,0 +1,8 @@
+import { partnersLogo } from '../../../../../../assets/index';
+import { PartnersPhoto } from './PhotoPartners.style';
+
+function PhotoPartners() {
+  return <PartnersPhoto src={partnersLogo} alt="Nossos parceiros" />;
+}
+
+export default PhotoPartners;
