+++ src/pages/Home/components/Partners/components/PhotoPartners/PhotoPartners.style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+/*Empreas que nos apoaiam */
+export const PartnersPhoto = styled.img`
+  position: absolute;
+  width: 367px;
+  height: 147.3px;
+  left: 536px;
+  top: 1900px;
+
+  background: url(image.png);
+`;
