+++ src/pages/Home/components/Partners/components/index.js
@@ -1,0 +1,4 @@
+import ButtonPartners from './ButtonPartners/ButtonPartners';
+import PhotoPartners from './PhotoPartners/PhotoPartners';
+
+export { ButtonPartners, PhotoPartners };
