+++ src/pages/Home/components/WhereAreWe/WhereAreWe.jsx
@@ -1,0 +1,13 @@
+import { WrapperTextWhere } from './components';
+import { DivWhere, MapImag } from './WhereAreWe.style';
+
+function WhereAreWe() {
+  return (
+    <DivWhere>
+      <WrapperTextWhere />
+      <MapImag />
+    </DivWhere>
+  );
+}
+
+export default WhereAreWe;
