+++ src/pages/Home/components/WhereAreWe/WhereAreWe.style.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const DivWhere = styled.menu`
+  position: absolute;
+  width: 1213px;
+  height: 546.27px;
+  left: 132px;
+  top: 100.55px;
+`;
+
+export const MapImag = styled.img`
+  margin-top: 5px;
+
+  position: absolute;
+  width: 100%;
+  max-width: 477px;
+  height: 530px;
+  left: 835px;
+  top: 1000px; /*232px */
+`;
