+++ src/pages/Home/components/WhereAreWe/components/WrapperTextWhere/WrapperTextWhere.style.js
@@ -1,0 +1,81 @@
+import styled from 'styled-components';
+
+export const OndeEstamosTitle = styled.p`
+  position: absolute;
+  width: 298px;
+  height: 60px;
+  left: 132px;
+  top: 1000px;
+
+  /* Headers/H2 */
+
+  font-family: Poppins;
+  font-style: normal;
+  font-weight: 600;
+  font-size: 40px;
+  line-height: 60px;
+  /* identical to box height */
+
+  /* Cores secundárias/Principal */
+
+  color: #e9e9e9;
+`;
+
+export const OndeEstamosSubTitle1 = styled.p`
+  position: absolute;
+  width: 61px;
+  height: 36px;
+  left: 132px;
+  top: 1100px;
+
+  /* Font families/Medium */
+
+  font-family: Poppins;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 24px;
+  line-height: 36px;
+  /* identical to box height */
+
+  /* Cores secundárias/Principal */
+
+  color: #e9e9e9;
+`;
+export const OndeEstamosSubTitle2 = styled.p`
+  position: absolute;
+  width: 223px;
+  height: 36px;
+  left: 132px;
+  top: 1150px;
+
+  /* Font families/Regular */
+
+  font-family: Poppins;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 24px;
+  line-height: 36px;
+  /* identical to box height */
+  /* Cores primárias/Principal */
+
+  color: #0095f6;
+`;
+
+export const TextWhere = styled.p`
+  position: absolute;
+  width: 485px;
+  height: 35px;
+  left: 132px;
+  top: 1230px;
+
+  /* Font families/Medium */
+  font-family: Poppins;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 24px;
+  line-height: 36px;
+  /* identical to box height */
+  /* Cores secundárias/Principal */
+
+  color: #e9e9e9;
+`;
