+++ src/pages/Home/components/WhereAreWe/components/index.js
@@ -1,0 +1,3 @@
+import WrapperTextWhere from './WrapperTextWhere/WrapperTextWhere';
+
+export { WrapperTextWhere };
