+++ src/pages/Home/components/index.js
@@ -1,0 +1,5 @@
+import AboutUs from './AboutUs/AboutUs';
+import Partners from './Partners/Partners';
+import WhereAreWe from './WhereAreWe/WhereAreWe';
+
+export { AboutUs, Partners, WhereAreWe };
